extracted expressCallback factory from the main index.js
