+++ src/controllers/index.js
@@ -2,10 +2,10 @@ const makeUsersEndpointHandler = require('./user-endpoint');
 const makeFriendsEndpointHandler = require('./friend-endpoint');
 const userActions = require('../domain/user-service');
 
-const handleUserRequest = makeUsersEndpointHandler({ userActions });
-const handleFriendRequest = makeFriendsEndpointHandler({ userActions });
+const userRequestHandler = makeUsersEndpointHandler({ userActions });
+const friendRequestHandler = makeFriendsEndpointHandler({ userActions });
 
 module.exports = {
-  handleUserRequest,
-  handleFriendRequest
+  userRequestHandler,
+  friendRequestHandler
 }
